pyproject.toml
[project]
name = "flask-live-transcription"
version = "0.1.0"
description = ""
authors = [{ name = "Jacob-Lasky", email = "jacob.s.lasky@gmail.com" }]
requires-python = ">=3.12, <3.13"
readme = "README.md"
dependencies = [
    "deepgram-sdk==6.0.1",
    "flask==3.0.0",
    "flask-socketio==5.3.6",
    "python-dotenv==1.0.0",
    "requests>=2.32.3,<3",
    "sounddevice>=0.5.2,<0.6",
    "pydub>=0.25.1,<0.26",
    "websocket-client>=1.8.0",
    "gevent>=23.0.0",
    "gunicorn>=21.0.0",
    "gevent-websocket>=0.10.1",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
